+++ CHANGELOG.md
@@ -7,10 +7,39 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## Unreleased
 
+### BREAKING CHANGES
+
+- `ImagePolicy.Spec.ImagePullSecrets` is now under
+  `ImagePolicy.Spec.ContainerRegistry.ImagePullSecrets`.
+
+Example Before:
+
+```
+apiVersion: hlnr.io/v1alpha1
+kind: ImagePolicy
+spec:
+  imagePullSecrets:
+    - name: docker-registry
+```
+
+After:
+
+```
+apiVersion: hlnr.io/v1alpha1
+kind: ImagePolicy
+spec:
+  containerRegistry:
+    name: docker
+    imagePullSecrets:
+      - name: docker-registry
+```
+
 ### Added
 
 - Added a health check for the GitHub Callback Server.
 - Added logging to indicate GH Callback Server requests.
+- Added `ImagePolicy.Spec.ContainerRegistry` to specify which container registry
+  to pull the image from.
 
 ### Fixed